parallelize message processing

+++ src/blockstorage/blockstorage.cpp
@@ -601,7 +601,7 @@ bool FlushStateToDiskInternal(CValidationState &state,
     static int64_t nSizeAfterLastFlush = 0;
     // The cache is close to the limit. Try to flush and trim.
     bool fCacheCritical = ((mode == FLUSH_STATE_IF_NEEDED) && (cacheSize > nCoinCacheMaxSize * 0.995)) ||
-                          (cacheSize - nSizeAfterLastFlush > nMaxCacheIncreaseSinceLastFlush);
+                          (cacheSize - nSizeAfterLastFlush > (int64_t)nMaxCacheIncreaseSinceLastFlush);
     // It's been a while since we wrote the block index to disk. Do this frequently, so we don't need to redownload
     // after a crash.
     bool fPeriodicWrite =
